requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
schedule>=1.2.1
aiosqlite>=0.19.0
greenlet>=3.0.3